refactor: split reducer into smaller parts
